requirements.txt
# .xlxs file analysis
pandas

# Document generation
python-docx
openpyxl